Modification of the GUI program to allow flexibility
